+++ src/main/java/loss/AbstractLossFunc.java
@@ -6,12 +6,12 @@ import java.io.Serializable;
 
 public abstract class AbstractLossFunc implements Serializable
 {
-    public static double EPSILON = 1e-15;
+    public static float EPSILON = 1e-15f;
 
-    public abstract double compute(Neuron [] predicted, double [] target);
-    public abstract void derive(Neuron [] predicted, double [] target);
+    public abstract float compute(Neuron [] predicted, float [] target);
+    public abstract void derive(Neuron [] predicted, float [] target);
 
-    protected void validate(Neuron[] predicted, double [] target) {
+    protected void validate(Neuron[] predicted, float [] target) {
         if(predicted.length != target.length) {
             throw new IllegalArgumentException("Mismatched array sizes");
         }
