+++ src/main/java/regularization/AlphaDropout.java
@@ -6,18 +6,18 @@ import java.util.Random;
 
 public class AlphaDropout implements Regulator
 {
-    private static final double saturation = -1.758099;
-    private final double q;
-    private final double a;
-    private final double b;
+    private static final float saturation = -1.758099f;
+    private final float q;
+    private final float a;
+    private final float b;
     private final boolean [] dropoutMask;
     private final Random rand = new Random(42);
 
-    public AlphaDropout(int size, double p) {
+    public AlphaDropout(int size, float p) {
         dropoutMask = new boolean[size];
-        q = 1.0 - p;
-        a = Math.pow((q + Math.pow(saturation, 2)*q*(1.0 - q)), -0.5);
-        b = -Math.pow((q + Math.pow(saturation, 2)*q*(1.0 - q)), -0.5)*((1.0 - q)*saturation);
+        q = 1.0f - p;
+        a = (float)Math.pow((q + Math.pow(saturation, 2)*q*(1.0f - q)), -0.5f);
+        b = (float)-Math.pow((q + Math.pow(saturation, 2)*q*(1.0f - q)), -0.5f)*((1.0f - q)*saturation);
     }
 
     @Override
@@ -26,19 +26,19 @@ public class AlphaDropout implements Regulator
         if(b == 1.0) return;
 
         for(int i = 0; i < dropoutMask.length; i++) {
-            double r = rand.nextDouble(1.0);
-            double d = q <= r ? 0.0 : 1.0;
+            float r = rand.nextFloat(1.0f);
+            float d = q <= r ? 0.0f : 1.0f;
 
-            double value = neuron[i].getValue();
-            neuron[i].setValue(a * (value * d + saturation * (1.0 - d)) + b);
+            float value = neuron[i].getValue();
+            neuron[i].setValue(a * (value * d + saturation * (1.0f - d)) + b);
 
-            dropoutMask[i] = (d == 0.0);
+            dropoutMask[i] = (d == 0.0f);
         }
     }
 
-    public double derive(int i) {
-        if(q == 1.0) return 1.0;
+    public float derive(int i) {
+        if(q == 1.0) return 1.0f;
 
-        return dropoutMask[i] ? 0.0 : a;
+        return dropoutMask[i] ? 0.0f : a;
     }
 }
